Fix startup errors (imports, etc.)
